refactor(converter): rename format_funcs.py to markdown_format_funcs.py
Clarifies that this module is Markdown-specific; Word and LaTeX
converters do not depend on it.  Update all three import sites
and keep git history via git mv.
